Add Score Label, Alert Message, Game Length
Added a state variable to keep track of total
score. This var is displayed in a label at bottom
of screen. A more descriptive alert message now
appears when a user gets a flag wrong and also
displays the score each time. Finally, the length
of the game is now 8 turns, in which at the end
an alert is shown with their accumulated score
(max of 8).