RPS NPC prediction and win/lose logic
Since the UI is finished, I implemented like 4
alerts to notify the user they haven't made a
prediction, result of match, result of game, and
prompt for the user to choose if they plan to win
or lose. Anyways, the game is over. What I'm
getting from this is SwiftUI is just a lot "looser"
I guess than UIKit. I still have some getting used
to.