Explore Date, Date Components, DatePicker, CoreML
I've used dates and date components in previous
projects (UIKIT) but it was a nice refresher. I'll
probably still have to look up or go back to
previous projects to solidify my understading, but
repetition is good for learning. However, it was
my first time using Apple's Create ML which I
didn't even know existed. It is really cool. I
hope to continue using it in future projects.