Change UI and implement CoreML logic
Used the model we create in CreateML to use the
inputs from the user (wakeup time, cups of coffee,
desired hours of sleep) to produce a bedtime for
the user. Used date, date components, steppers,
and date pickers for getting user input. Also,
changed UI by wrapping everything in a form and
combining sections using a VStack.