Create sections, Use picker, Dynamic Sleep Time
Instead of using vstacks to chunk together
relevant parts, I used sections within the form
and in my opinion it looks a bit cleaner. Another
task was to use a picker instead of a stepper for
the amount of coffee consumed. Personally I
thought the stepper was better, but it's good
practice anyways. Finally, instead of pressing
calculate every time, the bed time is dynamic and
shows a new time every time the user changes an
input. This is much better than having a button
and an alert.