Word Checker Functions, Display Error Messages
Created 3 new methods: isReal, isPossible, and
isOriginal. They check if the provided word is
even real by using UITextChecker, check if the
word is possible given letters from the root word,
and if the word was already used respectively. A
new method wordError was created to provide
error title and message to the alert for when
the word checking methods return false