Animation Experimentation
Experimenting with animations on a simple button.
Made an overlay of the button with an animation
that fades (opacity modifier) as it expands (scale
effect) with an easeOut animation that repeats
forever. The animationAmount onAppear starts at 2
so it the circle expands to double its size. The
.stroke(.red) modifier makes it so it is just the
outline of the circle.