Habit Tracker Play/Pause and Bank Functionality
First of all, create a LargeProgressView to
differentiate from the smaller ones used in the
gallery view. Additionally created a separate
timer view just so the UI is cleaner. Created
several binding variables to keep track of bank
time and share it across multiple timers. I want
the user to be able to choose between which habit
they want to dump their bank time into.