Storing Habits, New TimerView and EditHabitView
Currently using UserDefaults to store habits and
had to make Habits comfort to encodable, decodable
in the process. However, it is still not saving
properly so I have to check up on that.
Added a LargeTimerView (you can't spend time bank
from this one and it goes on the main screen) as
well as an EditHabit that esssentially works the
same as AddHabit, just modifying and saving the
habit.