Debug Saving to User Defaults
The habit array within content view only made a
copy of the habit array within allHabits, so the
didset property was never triggered. To combat
this, I made AllHabits and ObservableObject with
a published member habits so we are also notified
when the habits array is changed. Made allHabits a
@StateObject within content view because allHabits
is now an observable object.