Use Async Await to Pull Data and Images, Disabling
Use async methods to signify that the method will
take time and can be put to sleep. Use await when
calling those methods to signify sleep may occur.
We use these so that loading things from the
internet doesn't interupt the UI of our app.
Also learned .disable modifier on forms to disable
buttons or other things that require user input so
that we can only enable when a certain condition
has been met.