Add CheckoutView, Order Network Request
After providing a form for user to fill out
order details, order is converted to JSON and
sent to a server [(https://reqres.in/api/cupcakes)
is a website that sends back data to simulate
server requests] and we then convert that response
from JSON to Order again. Used CodingKeys to get
rid of underscore naming conentions.