Disable Buttons & Sliders, Adding More Filters
After finishing the project, I completed challenges
involving disabling the sliders and share button,
making more sliders for different filter keys, and
adding more filters.

I used 4 new state variables to keep track of
whether or not a filter/picture is picked so i can
enable sharing/sliders. Which sliders can be used
which is contingent on which filter is selected
and their respective keys. Finally, added more
filters to the confirmation dialog.