Adding Accessibility to Past Projects
In GuessTheFlag, added label for each flag
describing what it looks like, rather than just
saying which flag it is.
In WordScramble, created an accessibilityElement
on the list of words created so every child could
be read with the word and word count.
In Bookworm, for rating w/ stars, also created an
accessibilityElement on the stars, made the label
the name, the value the number of stars, and added
adjustableActions (swiping up/down) to change the
value of the stars.