MapKit Integration with iFace
ContactDetailView now provides a link to a map
view (when possible) of the location of the photo
of each person. I achieved this by converting the
image data from PhotosPicker to CIImage and
pulling image properties (which includes location
data) from the image. Used a Marker to put a pin
on the map as to where the image was taken.