List Selection, Tabview Navigation
In order to showcase list selection, made a list
from array of strings. List has initializer var
'selection' which can take a type, or a set of a
type for multiple selection. Used an EditButton
for selection.
Tabview navigation is a little different than what
Paul showcased. Each Tab in a tabview is
initialized with "Tab()", 'value' being a var you
must initialize. This value can be anything, but
must match selection type when TabView is
initialized (according to my understanding...)